refactor z-indexer
Z-indexer has been massively restructured. It’s no longer really inefficient, and it is no longer one line. Also, there are a few comments to help with following the logic. This should now work in browsers that lack support for `console.table`, `console.dir`, and the CSS typed object model.