create script
This is fairly robust and likely won’t need much tweaking. The only flaw in this approach (for duplicating and replacing the content in-place) is that code may still be running on the page, and I cannot find a way to remove the event listeners from the window and document. Everything within those should have no event listeners.